+++ 3548/1/Gemfile
@@ -1,7 +1,9 @@
 source 'https://rubygems.org'
 
-  gem 'octokit', '~>4.14.0'
-  gem 'terminal-table', '~>1.8.0'
-  gem 'nokogiri', '~>1.10.0'
-  gem 'dotenv', '~>2.7.4'
+gem 'octokit'
 
+gem 'terminal-table'
+
+gem 'nokogiri'
+
+gem 'dotenv'
